pyproject.toml
[project]
name = "collector-v1"
version = "0.1.0"
description = "Collector V1 - FastAPI monolith skeleton"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name" }]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg2-binary>=2.9",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.1",
  "bcrypt>=4.0",
  "python-json-logger==2.0.7",
  "ecdsa==0.19.1",
]


[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "httpx>=0.27",
  "ruff>=0.6",
  "pip-audit>=2.7",
  "bandit>=1.7.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"
fix = false

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "B008"] # on gère la longueur via line-length sans forcer E501

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
